store all temporary hardcoded data into one file
